+++ distributed/tests/test_scheduler.py
@@ -4159,23 +4159,6 @@ async def test_non_idempotent_plugins(s):
     assert s.plugins["nonidempotentplugin"].instance == "second"
 
 
-@gen_cluster(nthreads=[])
-async def test_plugin_removal(s):
-    class Plugin(SchedulerPlugin):
-        def __init__(self):
-            self.name = "plugin"
-
-    plugin = Plugin()
-    await s.register_scheduler_plugin(plugin=dumps(plugin))
-    assert "plugin" in s.plugins
-
-    await s.unregister_scheduler_plugin(name="plugin")
-    assert "plugin" not in s.plugins
-
-    with pytest.raises(ValueError, match="Could not find plugin"):
-        await s.unregister_scheduler_plugin(name="plugin")
-
-
 @gen_cluster(nthreads=[("", 1)])
 async def test_repr(s, a):
     async with Worker(s.address, nthreads=2) as b:  # name = address by default
